[Bronze III] Title: 알고리즘 수업 - 알고리즘의 수행 시간 3, Time: 184 ms, Memory: 17676 KB -BaekjoonHub

+++ 백준/Bronze/24264. 알고리즘 수업 － 알고리즘의 수행 시간 3/알고리즘 수업 － 알고리즘의 수행 시간 3.java
@@ -3,8 +3,8 @@ import java.util.Scanner;
 public class Main{
   public static void main(String[] args){
     Scanner sc = new Scanner(System.in);
-    long n = sc.nextLong();
-    System.out.println(n*n);
+    int n = sc.nextInt();
+    System.out.println((long)n*n);
     System.out.println(2);
   }
 }
